Split demo charts from more complicated example
